% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/browse.R
\name{browse}
\alias{browse}
\title{Browse a dataset webpage.}
\usage{
browse(x, ...)
}
\arguments{
\item{x}{datasetid or an object associated with a datasetid such \code{\link{info}}, \code{\link{griddap}} or \code{\link{tabledap}}}

\item{...}{Further args passed on to \code{\link[httr]{BROWSE}} (must be a named parameter)}

\item{url}{A URL for an ERDDAP server. Default: \url{http://upwell.pfeg.noaa.gov/erddap/}}
}
\value{
the value returned by \code{\link[httr]{BROWSE}}
}
\description{
Note that it is an error to call this when \code{base::interactive()} returns FALSE
}
\examples{
\dontrun{
 ed_browse('noaa_esrl_28d5_ac3a_bb06')
 }
}

