% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_delete.R
\name{cache_delete}
\alias{cache_delete}
\alias{cache_delete_all}
\title{Delete cached files}
\usage{
cache_delete(x, force = FALSE)

cache_delete_all(force = FALSE)
}
\arguments{
\item{x}{File names}

\item{force}{(logical) Should files be force deleted? Default: \code{FALSE}}
}
\description{
Delete cached files
}
\examples{
\dontrun{
# delete files by name in cache
# cache_delete('9911750294a039b8b517c8bf288978ea.csv')
# cache_delete(c('9911750294a039b8b517c8bf288978ea.csv',
#                'b26825b6737da13d6a52c28c8dfe690f.csv'))

# You can delete from the output of griddap or tabledap fxns
## tabledap
(table_res <- tabledap('erdCinpKfmBT'))
cache_delete(table_res)

## griddap
(out <- info('erdQMekm14day'))
(grid_res <- griddap(out,
 time = c('2015-12-28','2016-01-01'),
 latitude = c(24, 23),
 longitude = c(88, 90)
))
cache_delete(grid_res)
}
}
\seealso{
Other cache: 
\code{\link{cache_details}()},
\code{\link{cache_list}()},
\code{\link{cache_setup}()}
}
\concept{cache}
