% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{colors}
\alias{colors}
\title{cmocean colors
The cmocean color palette by Kristen Thyng as implemented in the R package "oce"}
\format{An object of class \code{list} of length 13.}
\usage{
colors
}
\description{
str(colors)
List of 13
 $ viridis
 $ cdom
 $ chlorophyll
 $ density
 $ freesurface
 $ oxygen
 $ par
 $ phase
 $ salinity
 $ temperature
 $ turbidity
 $ velocity
 $ vorticity
}
\keyword{datasets}

