% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{disk}
\alias{disk}
\alias{memory}
\title{Options for saving ERDDAP datasets.}
\usage{
disk(path = NULL, overwrite = TRUE)

memory()
}
\arguments{
\item{path}{Path to store files in. A directory, not a file.
Default: the root cache path, see \code{\link{cache_setup}}}

\item{overwrite}{(logical) Overwrite an existing file of the same name?
Default: \code{TRUE}}
}
\description{
Options for saving ERDDAP datasets.
}
