% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default-url.R
\name{eurl}
\alias{eurl}
\title{Default ERDDAP server URL}
\usage{
eurl()
}
\description{
Default ERDDAP server URL
}
\details{
default url is https://upwell.pfeg.noaa.gov/erddap/

You can set a default using an environment variable so you
don't have to pass anything to the URL parameter in your
function calls.

In your .Renviron file or similar set a URL for the environment
variable \code{RERDDAP_DEFAULT_URL}, like
\code{RERDDAP_DEFAULT_URL=https://upwell.pfeg.noaa.gov/erddap/}

It's important that you include a trailing slash in your URL
}
\examples{
eurl()
Sys.setenv(RERDDAP_DEFAULT_URL = "https://google.com")
Sys.getenv("RERDDAP_DEFAULT_URL")
eurl()
Sys.unsetenv("RERDDAP_DEFAULT_URL")
eurl()
}
