% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fipscounty}
\alias{fipscounty}
\title{Convert a FIPS County Code to/from a County Name}
\usage{
fipscounty(county = NULL, code = NULL,
  url = "http://coastwatch.pfeg.noaa.gov", ...)
}
\arguments{
\item{county}{character; A county name.}

\item{code}{numeric; A FIPS code.}

\item{url}{Base URL of the ERDDAP server}

\item{...}{Curl args passed on to \code{\link[httr]{GET}}}
}
\description{
Convert a FIPS County Code to/from a County Name
}
\examples{
\donttest{
fipscounty(code = "06053")
fipscounty(county = "CA, Monterey")
fipscounty(county = "OR, Multnomah")
}
}

