% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/info.R
\name{info}
\alias{as.info}
\alias{info}
\title{Get information on an ERDDAP dataset.}
\usage{
info(datasetid, ...)

as.info(x)
}
\arguments{
\item{datasetid}{Dataset id}

\item{...}{Further args passed on to \code{\link[httr]{GET}} (must be a named parameter)}

\item{x}{A datasetid or the output of \code{info}}
}
\value{
Prints a summary of the data on return, but you can index to various information.

The data is a list of length two with:
\itemize{
 \item variables - Data.frame of variables and their types
 \item alldata - List of data variables and their full attributes
}
Where \code{alldata} element has many data.frame's, one for each variable, with metadata
for that variable. E.g., for griddap dataset \code{noaa_pfeg_696e_ec99_6fa6}, \code{alldata}
has:
\itemize{
 \item NC_GLOBAL
 \item time
 \item latitude
 \item longitude
 \item sss
}
}
\description{
Get information on an ERDDAP dataset.
}
\examples{
\dontrun{
# grid dap datasets
info('noaa_pfeg_696e_ec99_6fa6')
info('noaa_ngdc_34bf_a95c_7e28')

(out <- erddap_search(query='temperature'))
info(out$info$dataset_id[5])
info(out$info$dataset_id[15])
info(out$info$dataset_id[25])
info(out$info$dataset_id[33])
info(out$info$dataset_id[65])
info(out$info$dataset_id[150])
info(out$info$dataset_id[400])
info(out$info$dataset_id[678])

out <- info(datasetid='noaa_ngdc_34bf_a95c_7e28')
## See brief overview of the variables and range of possible values, if given
out$variables
## all information on longitude
out$alldata$longitude
## all information on Climatological_Temperature
out$alldata$Climatological_Temperature

# table dap datasets
(out <- erddap_search(query='temperature', which = "table"))
info(out$info$dataset_id[1])
info(out$info$dataset_id[2])
info(out$info$dataset_id[3])
info(out$info$dataset_id[4])
info(out$info$dataset_id[54])

info(datasetid='erdCalCOFIfshsiz')
out <- info(datasetid='erdCinpKfmBT')
## See brief overview of the variables and range of possible values, if given
out$variables
## all information on longitude
out$alldata$longitude
## all information on Haliotis_corrugata_Mean_Density
out$alldata$Haliotis_corrugata_Mean_Density
}
}
\author{
Scott Chamberlain <myrmecocystus@gmail.com>
}
\references{
\url{http://upwell.pfeg.noaa.gov/erddap/index.html}
}

