% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/keywords.R, R/rerddap-package.r
\docType{data}
\name{keywords}
\alias{keywords}
\title{Convert a CF Standard Name to/from a GCMD Science Keyword}
\format{A character vector}
\usage{
keywords(cf = NULL, gcmd = NULL, url = "http://coastwatch.pfeg.noaa.gov",
  ...)
}
\arguments{
\item{cf}{character; A cf standard name
\url{http://cfconventions.org/Data/cf-standard-names/27/build/cf-standard-name-table.html}}

\item{gcmd}{character; A GCMD science keyword
\url{http://gcmd.gsfc.nasa.gov/learn/keyword_list.html}}

\item{url}{Base URL of the ERDDAP server}

\item{...}{Curl args passed on to \code{\link[httr]{GET}}}
}
\description{
Convert a CF Standard Name to/from a GCMD Science Keyword

keywords
}
\examples{
\donttest{
keywords(cf = "air_pressure")
cat(keywords(cf = "air_pressure"))
keywords(gcmd = "Atmosphere > Atmospheric Pressure > Sea Level Pressure")
cat(keywords(gcmd = "Atmosphere > Atmospheric Pressure > Sea Level Pressure"))
}
}
\keyword{datasets}

