% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotdap.R
\name{plotdap}
\alias{plotdap}
\alias{add_tabledap}
\alias{add_griddap}
\alias{add_ggplot}
\title{Visualize rerddap data}
\usage{
plotdap(method = c("ggplot2", "base"), mapData = maps::map("world", plot =
  FALSE, fill = TRUE), crs = NULL, datum = sf::st_crs(4326),
  mapTitle = NULL, mapFill = "gray80", mapColor = "gray90", ...)

add_tabledap(plot, table, var, color = c("#132B43", "#56B1F7"), size = 1.5,
  shape = 19, animate = FALSE, ani.args = list(filename = "ani.gif"), ...)

add_griddap(plot, grid, var, fill = "viridis", maxpixels = 10000,
  time = mean, animate = FALSE, ani.args = list(filename = "ani.gif"),
  ...)

add_ggplot(plot, ...)
}
\arguments{
\item{method}{the plotting method. Currently ggplot2 and base plotting
are supported.}

\item{mapData}{an object coercable to an sf object via \code{st_as_sf()}.}

\item{crs}{a coordinate reference system: integer with the epsg code,
or character with proj4string.}

\item{datum}{crs that provides datum to use when generating graticules.
Set to \code{NULL} to hide the graticule.}

\item{mapTitle}{a title for the map.}

\item{mapFill}{fill used for the map.}

\item{mapColor}{color used to draw boundaries of the map.}

\item{...}{for \code{plotdap()}/\code{add_tabledap()}/\code{add_griddap()},
arguments passed along to \code{geom_sf()} (if \code{method='ggplot2'},
otherwise ignored). For \code{add_ggplot2()}, these arguments may be
any object of class "gg" (e.g, \code{theme()}, \code{scale_fill_gradientn()},
\code{geom_point()}, etc.)}

\item{plot}{a \link{plotdap} object.}

\item{table}{a \link{tabledap} object.}

\item{var}{a formula defining a variable, or function of variables to visualize.}

\item{color}{a character string of length 1 a name in \link{colors}
or a vector of color codes. This defines the colorscale used to encode values
of \code{var}.}

\item{size}{the size of the symbol.}

\item{shape}{the shape of the symbol. For valid options, see the 'pch' values
section on \link{points}. \code{plot(0:25, 0:25, pch = 0:25)} also gives a
quick visual of the majority of possibilities.}

\item{animate}{whether to animate over the \code{time} variable (if it exists).
Currently only implemented for \code{method='ggplot2'} and requires the
gganimate package.}

\item{ani.args}{arguments passed along to \code{gganimate()}.}

\item{grid}{a \link{griddap} object.}

\item{maxpixels}{integer > 0. Maximum number of cells to use for the plot.
If maxpixels < ncell(x), sampleRegular is used before plotting.
If gridded=TRUE maxpixels may be ignored to get a larger sample}

\item{time}{how to resolve multiple time frames. Choose one of the following:
\itemize{
  \item A function to apply to each observation at a particular location
  (\link{mean} is the default).
  \item A character string (of length 1) matching a time value.
}}

\item{animate}{whether to animate over the \code{time} variable (if it exists).
Currently only implemented for \code{method='ggplot2'} and requires the
gganimate package.}

\item{ani.args}{arguments passed along to \code{gganimate()}.}

\item{plot}{a plotdap object}

\item{...}{ggplot2 elements}
}
\description{
Visualize data returned from rerddap servers. Use \code{plotdap()} to initialize
a plot, specify the plotting method (specifically, 'base' or 'ggplot2'),
and set some global options/parameters. Then use \code{add_tabledap()}
and/or \code{add_griddap()} to add "layers" of actual data to be visualized.

Add ggplot2 elements to a plotdap object
}
\details{
The "ggplot2" method is slower than "base" (especially
for high-res grids/rasters), but is more flexible/extensible. Additional ggplot2
layers, as well as scale defaults, labels, theming, etc. may be modified via
the \code{add_ggplot()} function. See the mapping vignette for an introduction
and overview of rerrdap's visualization methods --
\code{browseVignettes(package = "rerddap")}.
}
\examples{

plotdap()
plotdap("base")

\dontrun{
# tabledap examples
sardines <- tabledap(
  'FRDCPSTrawlLHHaulCatch',
  fields = c('latitude',  'longitude', 'time', 'scientific_name', 'subsample_count'),
  'time>=2010-01-01', 'time<=2012-01-01', 'scientific_name="Sardinops sagax"'
 )

p <- plotdap()
add_tabledap(p, sardines, ~subsample_count)
add_tabledap(p, sardines, ~log(subsample_count))

# using base R plotting
p <- plotdap("base")
add_tabledap(p, sardines, ~subsample_count)

# robinson projection
p <- plotdap(crs = "+proj=robin")
add_tabledap(p, sardines, ~subsample_count)

# griddap examples
murSST <- griddap(
  'jplMURSST41', latitude = c(22, 51), longitude = c(-140, -105),
  time = c('last', 'last'), fields = 'analysed_sst'
 )
p <- plotdap(crs = "+proj=robin")
add_griddap(p, murSST, ~analysed_sst)

# layer tables on top of grids
plotdap("base") \%>\%
  add_griddap(murSST, ~sst) \%>\%
  add_tabledap(sardines, ~subsample_count)

# multiple time periods
wind <- griddap(
  'erdQMwindmday', time = c('2016-11-16', '2017-01-16'),
  latitude = c(30, 50), longitude = c(210, 240),
  fields = 'x_wind'
)
p <- plotdap("base", mapTitle = "Average wind over time")
add_griddap(p, wind, ~x_wind)

}

add_ggplot(
 plotdap(
   crs = "+proj=laea +y_0=0 +lon_0=155 +lat_0=-90 +ellps=WGS84 +no_defs",
   mapColor = "black"
 ),
 theme_bw()
)


}
\seealso{
\code{\link{tabledap}()}, \code{\link{griddap}()}
}
\author{
Carson Sievert
}
