% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotdap.R
\name{print.ggplotdap}
\alias{print.ggplotdap}
\title{Print a ggplot plotdap object}
\usage{
\method{print}{ggplotdap}(plot, ...)
}
\arguments{
\item{plot}{a ggplotdap object}

\item{...}{currently unused}
}
\description{
Print a ggplot plotdap object
}
