% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotdap.R
\name{print.plotdap}
\alias{print.plotdap}
\title{Print a plotdap object}
\usage{
\method{print}{plotdap}(plot, ...)
}
\arguments{
\item{plot}{a plotdap object}

\item{...}{currently unused}
}
\description{
Print a plotdap object
}
