% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_list.R
\name{cache_list}
\alias{cache_list}
\title{List cached files}
\usage{
cache_list(cache_path = "~/.rerddap")
}
\arguments{
\item{cache_path}{path to cached files}
}
\description{
List cached files
}
\examples{
\dontrun{
# list files in cache
cache_list()

# List info for files
# cache_details(files = "243b4b41e19444515986ccf9cafbb1e9.nc")
# cache_details(files = "476ea03d8d246d81f2de02bd40524adb.csv")
cache_details()

# delete files by name in cache
# cache_delete(files = '9911750294a039b8b517c8bf288978ea.csv')
# cache_delete(files = c('9911750294a039b8b517c8bf288978ea.csv',
#                  'b26825b6737da13d6a52c28c8dfe690f.csv'))
}
}
\author{
Scott Chamberlain <myrmecocystus@gmail.com>
}
\references{
\url{http://upwell.pfeg.noaa.gov/erddap/index.html}
}
\seealso{
\code{\link{cache_delete}}, \code{\link{cache_details}}
}

