% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_list.R
\name{cache_list}
\alias{cache_list}
\title{List cached files}
\usage{
cache_list(cache_path = "~/.rerddap")
}
\arguments{
\item{cache_path}{path to cached files}
}
\description{
List cached files
}
\examples{
\dontrun{
# list files in cache
(x <- cache_list())

# List info for files
cache_details(x$nc[1])
cache_details(x$csv[1])
cache_details()

# delete files by name in cache
# cache_delete(x$nc[1])
# cache_delete(x$nc[2:3])
}
}
\references{
\url{https://upwell.pfeg.noaa.gov/erddap/index.html}
}
\seealso{
\code{\link{cache_delete}}, \code{\link{cache_details}}
}
\author{
Scott Chamberlain <myrmecocystus@gmail.com>
}
