% Generated by roxygen2 (4.0.2): do not edit by hand
\name{convert_time}
\alias{convert_time}
\title{Convert a CF Standard Name to/from a GCMD Science Keyword}
\usage{
convert_time(n = NULL, isoTime = NULL,
  units = "seconds since 1970-01-01T00:00:00Z",
  url = "http://coastwatch.pfeg.noaa.gov", ...)
}
\arguments{
\item{n}{numeric; A unix time number
\url{http://cfconventions.org/Data/cf-standard-names/27/build/cf-standard-name-table.html}}

\item{isoTime}{character; A string time representation
\url{http://gcmd.gsfc.nasa.gov/learn/keyword_list.html}}

\item{units}{character; Units to return. Default: "seconds since 1970-01-01T00:00:00Z"}

\item{url}{Base URL of the ERDDAP server}

\item{...}{Curl args passed on to \code{\link[httr]{GET}}}
}
\description{
Convert a CF Standard Name to/from a GCMD Science Keyword
}
\examples{
\donttest{
convert_time(n = 473472000)
convert_time(isoTime = "1985-01-02T00:00:00Z")
}
}

