% Generated by roxygen2 (4.0.2): do not edit by hand
\name{version}
\alias{version}
\title{Get ERDDAP version}
\usage{
version(url = "http://coastwatch.pfeg.noaa.gov", ...)
}
\arguments{
\item{url}{Base URL of the ERDDAP server}

\item{...}{Curl args passed on to \code{\link[httr]{GET}}}
}
\description{
Get ERDDAP version
}
\examples{
\donttest{
version("http://coastwatch.pfeg.noaa.gov")
version("https://bluehub.jrc.ec.europa.eu")
version("http://erddap.marine.ie")
}
}

